requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly.express 